1440: Allow Consent API controllers to be configured for specific OB versions (#240)
* 1440: Allow Consent API controllers to be configured for specific OB versions

Refactoring the remaining Consent API payment controllers as per: https://github.com/SecureApiGateway/secure-api-gateway-ob-uk-rcs/pull/239

https://github.com/SecureApiGateway/SecureApiGateway/issues/1440
