+++ secure-api-gateway-ob-uk-rcs-consent-store/secure-api-gateway-ob-uk-rcs-consent-store-api/src/main/java/com/forgerock/sapi/gateway/rcs/consent/store/api/account/AccountAccessConsentApi.java
@@ -13,9 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.forgerock.sapi.gateway.rcs.consent.store.api.payment.vrp.v3_1_10;
-
-import jakarta.validation.Valid;
+package com.forgerock.sapi.gateway.rcs.consent.store.api.account;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
@@ -26,26 +24,24 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
 import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.RejectConsentRequest;
-import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.payment.AuthorisePaymentConsentRequest;
-import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.payment.vrp.v3_1_10.CreateDomesticVRPConsentRequest;
-import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.payment.vrp.v3_1_10.DomesticVRPConsent;
+import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.account.v3_1_10.AccountAccessConsent;
+import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.account.v3_1_10.AuthoriseAccountAccessConsentRequest;
+import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.account.v3_1_10.CreateAccountAccessConsentRequest;
 
-import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
+import jakarta.validation.Valid;
 import uk.org.openbanking.datamodel.v3.error.OBErrorResponse1;
 
 @Validated
-@Api(tags = {"v3.1.10"})
-@RequestMapping(value = "/consent/store/v3.1.10")
-public interface DomesticVRPConsentApi {
+public interface AccountAccessConsentApi {
 
-    @ApiOperation(value = "Create Domestic VRP Consent")
+    @ApiOperation(value = "Create Account Access Consent")
     @ApiResponses(value = {
-            @ApiResponse(code = 201, message = "DomesticVRPConsent object representing the consent created",
-                         response = DomesticVRPConsent.class),
+            @ApiResponse(code = 201, message = "AccountAccessConsent object representing the consent created",
+                    response = AccountAccessConsent.class),
             @ApiResponse(code = 400, message = "Bad request", response = OBErrorResponse1.class),
             @ApiResponse(code = 403, message = "Forbidden", response = OBErrorResponse1.class),
             @ApiResponse(code = 404, message = "Not found"),
@@ -53,19 +49,19 @@ public interface DomesticVRPConsentApi {
             @ApiResponse(code = 406, message = "Not Acceptable"),
             @ApiResponse(code = 500, message = "Internal Server Error", response = OBErrorResponse1.class)
     })
-    @RequestMapping(value = "/domestic-vrp-consents",
+    @RequestMapping(value = "/account-access-consents",
             consumes = {"application/json; charset=utf-8"},
             produces = {"application/json; charset=utf-8"},
             method = RequestMethod.POST)
-    ResponseEntity<DomesticVRPConsent> createConsent(@ApiParam(value = "Create Consent Request", required = true)
-                                                     @Valid
-                                                     @RequestBody CreateDomesticVRPConsentRequest request);
+    ResponseEntity<AccountAccessConsent> createConsent(@ApiParam(value = "Create Consent Request", required = true)
+                                                       @Valid
+                                                       @RequestBody CreateAccountAccessConsentRequest request);
 
 
-    @ApiOperation(value = "Get Domestic VRP Consent")
+    @ApiOperation(value = "Get Account Access Consent")
     @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "DomesticVRPConsent object representing the consent created",
-                         response = DomesticVRPConsent.class),
+            @ApiResponse(code = 200, message = "AccountAccessConsent object representing the consent created",
+                    response = AccountAccessConsent.class),
             @ApiResponse(code = 400, message = "Bad request", response = OBErrorResponse1.class),
             @ApiResponse(code = 403, message = "Forbidden", response = OBErrorResponse1.class),
             @ApiResponse(code = 404, message = "Not found"),
@@ -73,17 +69,17 @@ public interface DomesticVRPConsentApi {
             @ApiResponse(code = 406, message = "Not Acceptable"),
             @ApiResponse(code = 500, message = "Internal Server Error", response = OBErrorResponse1.class)
     })
-    @RequestMapping(value = "/domestic-vrp-consents/{consentId}",
+    @RequestMapping(value = "/account-access-consents/{consentId}",
             produces = {"application/json; charset=utf-8"},
             method = RequestMethod.GET)
-    ResponseEntity<DomesticVRPConsent> getConsent(@PathVariable(value = "consentId") String consentId,
-                                                  @RequestHeader(value = "x-api-client-id") String apiClientId);
+    ResponseEntity<AccountAccessConsent> getConsent(@PathVariable(value = "consentId") String consentId,
+                                                    @RequestHeader(value = "x-api-client-id") String apiClientId);
 
 
-    @ApiOperation(value = "Authorise Domestic VRP Consent")
+    @ApiOperation(value = "Authorise Account Access Consent")
     @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "DomesticVRPConsent object representing the consent created",
-                         response = DomesticVRPConsent.class),
+            @ApiResponse(code = 200, message = "AccountAccessConsent object representing the consent created",
+                    response = AccountAccessConsent.class),
             @ApiResponse(code = 400, message = "Bad request", response = OBErrorResponse1.class),
             @ApiResponse(code = 403, message = "Forbidden", response = OBErrorResponse1.class),
             @ApiResponse(code = 404, message = "Not found"),
@@ -91,20 +87,20 @@ public interface DomesticVRPConsentApi {
             @ApiResponse(code = 406, message = "Not Acceptable"),
             @ApiResponse(code = 500, message = "Internal Server Error", response = OBErrorResponse1.class)
     })
-    @RequestMapping(value = "/domestic-vrp-consents/{consentId}/authorise",
+    @RequestMapping(value = "/account-access-consents/{consentId}/authorise",
             consumes = {"application/json; charset=utf-8"},
             produces = {"application/json; charset=utf-8"},
             method = RequestMethod.POST)
-    ResponseEntity<DomesticVRPConsent> authoriseConsent(@PathVariable(value = "consentId") String consentId,
-                                                        @ApiParam(value = "Authorise Consent Request", required = true)
-                                                        @Valid
-                                                        @RequestBody AuthorisePaymentConsentRequest request);
+    ResponseEntity<AccountAccessConsent> authoriseConsent(@PathVariable(value = "consentId") String consentId,
+                                                          @ApiParam(value = "Authorise Consent Request", required = true)
+                                                          @Valid
+                                                          @RequestBody AuthoriseAccountAccessConsentRequest request);
 
 
-    @ApiOperation(value = "Reject Domestic VRP Consent")
+    @ApiOperation(value = "Reject Account Access Consent")
     @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "DomesticVRPConsent object representing the consent created",
-                         response = DomesticVRPConsent.class),
+            @ApiResponse(code = 200, message = "AccountAccessConsent object representing the consent created",
+                    response = AccountAccessConsent.class),
             @ApiResponse(code = 400, message = "Bad request", response = OBErrorResponse1.class),
             @ApiResponse(code = 403, message = "Forbidden", response = OBErrorResponse1.class),
             @ApiResponse(code = 404, message = "Not found"),
@@ -112,17 +108,17 @@ public interface DomesticVRPConsentApi {
             @ApiResponse(code = 406, message = "Not Acceptable"),
             @ApiResponse(code = 500, message = "Internal Server Error", response = OBErrorResponse1.class)
     })
-    @RequestMapping(value = "/domestic-vrp-consents/{consentId}/reject",
+    @RequestMapping(value = "/account-access-consents/{consentId}/reject",
             consumes = {"application/json; charset=utf-8"},
             produces = {"application/json; charset=utf-8"},
             method = RequestMethod.POST)
-    ResponseEntity<DomesticVRPConsent> rejectConsent(@PathVariable(value = "consentId") String consentId,
-                                                     @ApiParam(value = "Reject Consent Request", required = true)
-                                                     @Valid
-                                                     @RequestBody RejectConsentRequest request);
+    ResponseEntity<AccountAccessConsent> rejectConsent(@PathVariable(value = "consentId") String consentId,
+                                                       @ApiParam(value = "Reject Consent Request", required = true)
+                                                       @Valid
+                                                       @RequestBody RejectConsentRequest request);
 
 
-    @ApiOperation(value = "Delete Domestic VRP Consent")
+    @ApiOperation(value = "Delete Account Access Consent")
     @ApiResponses(value = {
             @ApiResponse(code = 200, message = "Delete successful"),
             @ApiResponse(code = 400, message = "Bad request", response = OBErrorResponse1.class),
@@ -132,9 +128,10 @@ public interface DomesticVRPConsentApi {
             @ApiResponse(code = 406, message = "Not Acceptable"),
             @ApiResponse(code = 500, message = "Internal Server Error", response = OBErrorResponse1.class)
     })
-    @RequestMapping(value = "/domestic-vrp-consents/{consentId}",
+    @RequestMapping(value = "/account-access-consents/{consentId}",
             method = RequestMethod.DELETE)
     ResponseEntity<Void> deleteConsent(@PathVariable(value = "consentId") String consentId,
                                        @RequestHeader(value = "x-api-client-id") String apiClientId);
+
 
 }
