+++ secure-api-gateway-ob-uk-rcs-consent-store/secure-api-gateway-ob-uk-rcs-consent-store-api/src/main/java/com/forgerock/sapi/gateway/rcs/consent/store/api/customerinfo/CustomerInfoConsentApi.java
@@ -13,9 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.forgerock.sapi.gateway.rcs.consent.store.api.payment.internationalstandingorder.v3_1_10;
-
-import jakarta.validation.Valid;
+package com.forgerock.sapi.gateway.rcs.consent.store.api.customerinfo;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
@@ -26,27 +24,27 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
 import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.RejectConsentRequest;
-import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.payment.AuthorisePaymentConsentRequest;
-import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.payment.ConsumePaymentConsentRequest;
-import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.payment.internationalstandingorder.v3_1_10.CreateInternationalStandingOrderConsentRequest;
-import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.payment.internationalstandingorder.v3_1_10.InternationalStandingOrderConsent;
+import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.customerinfo.v1_0.AuthoriseCustomerInfoConsentRequest;
+import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.customerinfo.v1_0.CreateCustomerInfoConsentRequest;
+import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.customerinfo.v1_0.CustomerInfoConsent;
 
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
+import jakarta.validation.Valid;
 import uk.org.openbanking.datamodel.v3.error.OBErrorResponse1;
 
 @Validated
-@Api(tags = {"v3.1.10"})
-@RequestMapping(value = "/consent/store/v3.1.10")
-public interface InternationalStandingOrderConsentApi {
+@Api(tags = {"v1.0"})
+@RequestMapping(value = "/consent/store/v1.0")
+public interface CustomerInfoConsentApi {
 
-    @ApiOperation(value = "Create International Standing Order Consent")
+    @ApiOperation(value = "Create Customer Info Consent")
     @ApiResponses(value = {
-            @ApiResponse(code = 201, message = "InternationalStandingOrderConsent object representing the consent created",
-                         response = InternationalStandingOrderConsent.class),
+            @ApiResponse(code = 201, message = "CustomerInfoConsent object representing the consent created",
+                    response = CustomerInfoConsent.class),
             @ApiResponse(code = 400, message = "Bad request", response = OBErrorResponse1.class),
             @ApiResponse(code = 403, message = "Forbidden", response = OBErrorResponse1.class),
             @ApiResponse(code = 404, message = "Not found"),
@@ -54,19 +52,19 @@ public interface InternationalStandingOrderConsentApi {
             @ApiResponse(code = 406, message = "Not Acceptable"),
             @ApiResponse(code = 500, message = "Internal Server Error", response = OBErrorResponse1.class)
     })
-    @RequestMapping(value = "/international-standing-order-consents",
+    @RequestMapping(value = "/customer-info-consents",
             consumes = {"application/json; charset=utf-8"},
             produces = {"application/json; charset=utf-8"},
             method = RequestMethod.POST)
-    ResponseEntity<InternationalStandingOrderConsent> createConsent(@ApiParam(value = "Create Consent Request", required = true)
-                                                                    @Valid
-                                                                    @RequestBody CreateInternationalStandingOrderConsentRequest request);
+    ResponseEntity<CustomerInfoConsent> createConsent(@ApiParam(value = "Create Consent Request", required = true)
+                                                      @Valid
+                                                      @RequestBody CreateCustomerInfoConsentRequest request);
 
 
-    @ApiOperation(value = "Get International Standing Order Consent")
+    @ApiOperation(value = "Get Customer Info Consent")
     @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "InternationalStandingOrderConsent object representing the consent created",
-                         response = InternationalStandingOrderConsent.class),
+            @ApiResponse(code = 200, message = "CustomerInfoConsent object representing the consent created",
+                    response = CustomerInfoConsent.class),
             @ApiResponse(code = 400, message = "Bad request", response = OBErrorResponse1.class),
             @ApiResponse(code = 403, message = "Forbidden", response = OBErrorResponse1.class),
             @ApiResponse(code = 404, message = "Not found"),
@@ -74,17 +72,17 @@ public interface InternationalStandingOrderConsentApi {
             @ApiResponse(code = 406, message = "Not Acceptable"),
             @ApiResponse(code = 500, message = "Internal Server Error", response = OBErrorResponse1.class)
     })
-    @RequestMapping(value = "/international-standing-order-consents/{consentId}",
+    @RequestMapping(value = "/customer-info-consents/{consentId}",
             produces = {"application/json; charset=utf-8"},
             method = RequestMethod.GET)
-    ResponseEntity<InternationalStandingOrderConsent> getConsent(@PathVariable(value = "consentId") String consentId,
-                                                                 @RequestHeader(value = "x-api-client-id") String apiClientId);
+    ResponseEntity<CustomerInfoConsent> getConsent(@PathVariable(value = "consentId") String consentId,
+                                                   @RequestHeader(value = "x-api-client-id") String apiClientId);
 
 
-    @ApiOperation(value = "Authorise International Standing Order Consent")
+    @ApiOperation(value = "Authorise Customer Info Consent")
     @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "InternationalStandingOrderConsent object representing the consent created",
-                         response = InternationalStandingOrderConsent.class),
+            @ApiResponse(code = 200, message = "CustomerInfoConsent object representing the consent created",
+                    response = CustomerInfoConsent.class),
             @ApiResponse(code = 400, message = "Bad request", response = OBErrorResponse1.class),
             @ApiResponse(code = 403, message = "Forbidden", response = OBErrorResponse1.class),
             @ApiResponse(code = 404, message = "Not found"),
@@ -92,20 +90,20 @@ public interface InternationalStandingOrderConsentApi {
             @ApiResponse(code = 406, message = "Not Acceptable"),
             @ApiResponse(code = 500, message = "Internal Server Error", response = OBErrorResponse1.class)
     })
-    @RequestMapping(value = "/international-standing-order-consents/{consentId}/authorise",
+    @RequestMapping(value = "/customer-info-consents/{consentId}/authorise",
             consumes = {"application/json; charset=utf-8"},
             produces = {"application/json; charset=utf-8"},
             method = RequestMethod.POST)
-    ResponseEntity<InternationalStandingOrderConsent> authoriseConsent(@PathVariable(value = "consentId") String consentId,
-                                                                       @ApiParam(value = "Authorise Consent Request", required = true)
-                                                                       @Valid
-                                                                       @RequestBody AuthorisePaymentConsentRequest request);
+    ResponseEntity<CustomerInfoConsent> authoriseConsent(@PathVariable(value = "consentId") String consentId,
+                                                         @ApiParam(value = "Authorise Consent Request", required = true)
+                                                         @Valid
+                                                         @RequestBody AuthoriseCustomerInfoConsentRequest request);
 
 
-    @ApiOperation(value = "Reject International Standing Order Consent")
+    @ApiOperation(value = "Reject Customer Info Consent")
     @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "InternationalStandingOrderConsent object representing the consent created",
-                         response = InternationalStandingOrderConsent.class),
+            @ApiResponse(code = 200, message = "CustomerInfoConsent object representing the consent created",
+                    response = CustomerInfoConsent.class),
             @ApiResponse(code = 400, message = "Bad request", response = OBErrorResponse1.class),
             @ApiResponse(code = 403, message = "Forbidden", response = OBErrorResponse1.class),
             @ApiResponse(code = 404, message = "Not found"),
@@ -113,21 +111,19 @@ public interface InternationalStandingOrderConsentApi {
             @ApiResponse(code = 406, message = "Not Acceptable"),
             @ApiResponse(code = 500, message = "Internal Server Error", response = OBErrorResponse1.class)
     })
-    @RequestMapping(value = "/international-standing-order-consents/{consentId}/reject",
+    @RequestMapping(value = "/customer-info-consents/{consentId}/reject",
             consumes = {"application/json; charset=utf-8"},
             produces = {"application/json; charset=utf-8"},
             method = RequestMethod.POST)
-    ResponseEntity<InternationalStandingOrderConsent> rejectConsent(@PathVariable(value = "consentId") String consentId,
-                                                                    @ApiParam(value = "Reject Consent Request", required = true)
-                                                                    @Valid
-                                                                    @RequestBody RejectConsentRequest request);
+    ResponseEntity<CustomerInfoConsent> rejectConsent(@PathVariable(value = "consentId") String consentId,
+                                                      @ApiParam(value = "Reject Consent Request", required = true)
+                                                      @Valid
+                                                      @RequestBody RejectConsentRequest request);
 
 
-
-    @ApiOperation(value = "Consume International Standing Order Consent")
+    @ApiOperation(value = "Delete Customer Info Consent")
     @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "InternationalStandingOrderConsent object representing the consent created",
-                         response = InternationalStandingOrderConsent.class),
+            @ApiResponse(code = 200, message = "Delete successful"),
             @ApiResponse(code = 400, message = "Bad request", response = OBErrorResponse1.class),
             @ApiResponse(code = 403, message = "Forbidden", response = OBErrorResponse1.class),
             @ApiResponse(code = 404, message = "Not found"),
@@ -135,13 +131,10 @@ public interface InternationalStandingOrderConsentApi {
             @ApiResponse(code = 406, message = "Not Acceptable"),
             @ApiResponse(code = 500, message = "Internal Server Error", response = OBErrorResponse1.class)
     })
-    @RequestMapping(value = "/international-standing-order-consents/{consentId}/consume",
-            consumes = {"application/json; charset=utf-8"},
-            produces = {"application/json; charset=utf-8"},
-            method = RequestMethod.POST)
-    ResponseEntity<InternationalStandingOrderConsent> consumeConsent(@PathVariable(value = "consentId") String consentId,
-                                                                     @ApiParam(value = "Consume Consent Request", required = true)
-                                                                     @Valid
-                                                                     @RequestBody ConsumePaymentConsentRequest request);
+    @RequestMapping(value = "/customer-info-consents/{consentId}",
+            method = RequestMethod.DELETE)
+    ResponseEntity<Void> deleteConsent(@PathVariable(value = "consentId") String consentId,
+                                       @RequestHeader(value = "x-api-client-id") String apiClientId);
+
 
 }
