+++ secure-api-gateway-ob-uk-rcs-consent-store/secure-api-gateway-ob-uk-rcs-consent-store-api/src/main/java/com/forgerock/sapi/gateway/rcs/consent/store/api/payment/domesticscheduled/BaseDomesticScheduledPaymentConsentApiController.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.forgerock.sapi.gateway.rcs.consent.store.api.payment.domestic;
+package com.forgerock.sapi.gateway.rcs.consent.store.api.payment.domesticscheduled;
 
 import static java.util.Objects.requireNonNull;
 
@@ -28,64 +28,67 @@ import org.springframework.http.ResponseEntity;
 import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.RejectConsentRequest;
 import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.payment.AuthorisePaymentConsentRequest;
 import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.payment.ConsumePaymentConsentRequest;
-import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.payment.domestic.v3_1_10.CreateDomesticPaymentConsentRequest;
-import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.payment.domestic.v3_1_10.DomesticPaymentConsent;
-import com.forgerock.sapi.gateway.rcs.consent.store.repo.entity.payment.domestic.DomesticPaymentConsentEntity;
+import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.payment.domesticscheduled.v3_1_10.CreateDomesticScheduledPaymentConsentRequest;
+import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.payment.domesticscheduled.v3_1_10.DomesticScheduledPaymentConsent;
+import com.forgerock.sapi.gateway.rcs.consent.store.repo.entity.payment.domestic.DomesticScheduledPaymentConsentEntity;
 import com.forgerock.sapi.gateway.rcs.consent.store.repo.service.payment.PaymentAuthoriseConsentArgs;
-import com.forgerock.sapi.gateway.rcs.consent.store.repo.service.payment.domestic.DomesticPaymentConsentService;
+import com.forgerock.sapi.gateway.rcs.consent.store.repo.service.payment.domestic.DomesticScheduledPaymentConsentService;
 import com.forgerock.sapi.gateway.uk.common.shared.api.meta.obie.OBVersion;
 
 import uk.org.openbanking.datamodel.v3.payment.OBPaymentConsentStatus;
 
 /**
- * Base implementation of {@link DomesticPaymentConsentApi} containing the logic that is common across versions.
+ * Base implementation of {@link DomesticScheduledPaymentConsentApi} containing the logic that is common across versions.
  * The {@link OBVersion} can be plugged in via the constructor.
  * <p>
  * Version specific controller implementations should extend this class and specify the following annotations:
  * {@link org.springframework.stereotype.Controller}, {@link org.springframework.web.bind.annotation.RequestMapping}
  * and {@link io.swagger.annotations.Api}.
  */
-public class BaseDomesticPaymentConsentApiController implements DomesticPaymentConsentApi {
+public class BaseDomesticScheduledPaymentConsentApiController implements DomesticScheduledPaymentConsentApi {
 
     private final Logger logger = LoggerFactory.getLogger(getClass());
 
-    private final DomesticPaymentConsentService consentService;
+    private final DomesticScheduledPaymentConsentService consentService;
     private final Supplier<DateTime> idempotencyKeyExpirationSupplier;
     private final OBVersion obVersion;
 
-    public BaseDomesticPaymentConsentApiController(DomesticPaymentConsentService consentService,
-                                                   Supplier<DateTime> idempotencyKeyExpirationSupplier,
-                                                   OBVersion obVersion) {
+    public BaseDomesticScheduledPaymentConsentApiController(DomesticScheduledPaymentConsentService consentService,
+                                                            Supplier<DateTime> idempotencyKeyExpirationSupplier,
+                                                            OBVersion obVersion) {
         this.consentService = requireNonNull(consentService, "consentService must be provided");
         this.idempotencyKeyExpirationSupplier = requireNonNull(idempotencyKeyExpirationSupplier, "idempotencyKeyExpirationSupplier must be provided");
         this.obVersion = requireNonNull(obVersion, "obVersion must be provided");
     }
 
     @Override
-    public ResponseEntity<DomesticPaymentConsent> createConsent(CreateDomesticPaymentConsentRequest request) {
+    public ResponseEntity<DomesticScheduledPaymentConsent> createConsent(CreateDomesticScheduledPaymentConsentRequest request) {
+
         logger.info("Attempting to createConsent: {}", request);
-        final DomesticPaymentConsentEntity domesticPaymentConsent = new DomesticPaymentConsentEntity();
-        domesticPaymentConsent.setRequestVersion(obVersion);
-        domesticPaymentConsent.setApiClientId(request.getApiClientId());
-        domesticPaymentConsent.setRequestObj(request.getConsentRequest());
-        domesticPaymentConsent.setStatus(OBPaymentConsentStatus.AWAITINGAUTHORISATION.toString());
-        domesticPaymentConsent.setCharges(request.getCharges());
-        domesticPaymentConsent.setIdempotencyKey(request.getIdempotencyKey());
-        domesticPaymentConsent.setIdempotencyKeyExpiration(idempotencyKeyExpirationSupplier.get());
-        final DomesticPaymentConsentEntity persistedEntity = consentService.createConsent(domesticPaymentConsent);
+        final DomesticScheduledPaymentConsentEntity domesticScheduledPaymentConsent = new DomesticScheduledPaymentConsentEntity();
+        domesticScheduledPaymentConsent.setRequestVersion(obVersion);
+        domesticScheduledPaymentConsent.setApiClientId(request.getApiClientId());
+        domesticScheduledPaymentConsent.setRequestObj(request.getConsentRequest());
+        domesticScheduledPaymentConsent.setStatus(OBPaymentConsentStatus.AWAITINGAUTHORISATION.toString());
+        domesticScheduledPaymentConsent.setCharges(request.getCharges());
+        domesticScheduledPaymentConsent.setIdempotencyKey(request.getIdempotencyKey());
+        domesticScheduledPaymentConsent.setIdempotencyKeyExpiration(idempotencyKeyExpirationSupplier.get());
+        final DomesticScheduledPaymentConsentEntity persistedEntity = consentService.createConsent(domesticScheduledPaymentConsent);
         logger.info("Consent created with id: {}", persistedEntity.getId());
 
         return new ResponseEntity<>(convertEntityToDto(persistedEntity), HttpStatus.CREATED);
     }
 
     @Override
-    public ResponseEntity<DomesticPaymentConsent> getConsent(String consentId, String apiClientId) {
+    public ResponseEntity<DomesticScheduledPaymentConsent> getConsent(String consentId, String apiClientId) {
         logger.info("Attempting to getConsent - id: {}, for apiClientId: {}", consentId, apiClientId);
         return ResponseEntity.ok(convertEntityToDto(consentService.getConsent(consentId, apiClientId)));
     }
 
     @Override
-    public ResponseEntity<DomesticPaymentConsent> authoriseConsent(String consentId, AuthorisePaymentConsentRequest request) {
+    public ResponseEntity<DomesticScheduledPaymentConsent> authoriseConsent(String consentId,
+            AuthorisePaymentConsentRequest request) {
+
         logger.info("Attempting to authoriseConsent - id: {}, request: {}", consentId, request);
         final PaymentAuthoriseConsentArgs paymentAuthoriseConsentArgs = new PaymentAuthoriseConsentArgs(consentId,
                 request.getApiClientId(), request.getResourceOwnerId(), request.getAuthorisedDebtorAccountId());
@@ -94,19 +97,19 @@ public class BaseDomesticPaymentConsentApiController implements DomesticPaymentC
     }
 
     @Override
-    public ResponseEntity<DomesticPaymentConsent> rejectConsent(String consentId, RejectConsentRequest request) {
+    public ResponseEntity<DomesticScheduledPaymentConsent> rejectConsent(String consentId, RejectConsentRequest request) {
         logger.info("Attempting to rejectConsent - id: {}, request: {}", consentId, request);
         return ResponseEntity.ok(convertEntityToDto(consentService.rejectConsent(consentId, request.getApiClientId(), request.getResourceOwnerId())));
     }
 
     @Override
-    public ResponseEntity<DomesticPaymentConsent> consumeConsent(String consentId, ConsumePaymentConsentRequest request) {
+    public ResponseEntity<DomesticScheduledPaymentConsent> consumeConsent(String consentId, ConsumePaymentConsentRequest request) {
         logger.info("Attempting to consumeConsent - id: {}, request: {}", consentId, request);
         return ResponseEntity.ok(convertEntityToDto(consentService.consumeConsent(consentId, request.getApiClientId())));
     }
 
-    private DomesticPaymentConsent convertEntityToDto(DomesticPaymentConsentEntity entity) {
-        final DomesticPaymentConsent dto = new DomesticPaymentConsent();
+    private DomesticScheduledPaymentConsent convertEntityToDto(DomesticScheduledPaymentConsentEntity entity) {
+        final DomesticScheduledPaymentConsent dto = new DomesticScheduledPaymentConsent();
         dto.setId(entity.getId());
         dto.setStatus(entity.getStatus());
         dto.setRequestObj(entity.getRequestObj());
