+++ secure-api-gateway-ob-uk-rcs-consent-store/secure-api-gateway-ob-uk-rcs-consent-store-api/src/main/java/com/forgerock/sapi/gateway/rcs/consent/store/api/payment/internationalstandingorder/InternationalStandingOrderConsentApi.java
@@ -13,9 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.forgerock.sapi.gateway.rcs.consent.store.api.account.v3_1_10;
-
-import jakarta.validation.Valid;
+package com.forgerock.sapi.gateway.rcs.consent.store.api.payment.internationalstandingorder;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
@@ -26,27 +24,25 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
 import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.RejectConsentRequest;
-import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.account.v3_1_10.AccountAccessConsent;
-import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.account.v3_1_10.AuthoriseAccountAccessConsentRequest;
-import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.account.v3_1_10.CreateAccountAccessConsentRequest;
+import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.payment.AuthorisePaymentConsentRequest;
+import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.payment.ConsumePaymentConsentRequest;
+import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.payment.internationalstandingorder.v3_1_10.CreateInternationalStandingOrderConsentRequest;
+import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.payment.internationalstandingorder.v3_1_10.InternationalStandingOrderConsent;
 
-import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
+import jakarta.validation.Valid;
 import uk.org.openbanking.datamodel.v3.error.OBErrorResponse1;
 
 @Validated
-@Api(tags = {"v3.1.10"})
-@RequestMapping(value = "/consent/store/v3.1.10")
-public interface AccountAccessConsentApi {
+public interface InternationalStandingOrderConsentApi {
 
-
-    @ApiOperation(value = "Create Account Access Consent")
+    @ApiOperation(value = "Create International Standing Order Consent")
     @ApiResponses(value = {
-            @ApiResponse(code = 201, message = "AccountAccessConsent object representing the consent created",
-                    response = AccountAccessConsent.class),
+            @ApiResponse(code = 201, message = "InternationalStandingOrderConsent object representing the consent created",
+                         response = InternationalStandingOrderConsent.class),
             @ApiResponse(code = 400, message = "Bad request", response = OBErrorResponse1.class),
             @ApiResponse(code = 403, message = "Forbidden", response = OBErrorResponse1.class),
             @ApiResponse(code = 404, message = "Not found"),
@@ -54,19 +50,19 @@ public interface AccountAccessConsentApi {
             @ApiResponse(code = 406, message = "Not Acceptable"),
             @ApiResponse(code = 500, message = "Internal Server Error", response = OBErrorResponse1.class)
     })
-    @RequestMapping(value = "/account-access-consents",
+    @RequestMapping(value = "/international-standing-order-consents",
             consumes = {"application/json; charset=utf-8"},
             produces = {"application/json; charset=utf-8"},
             method = RequestMethod.POST)
-    ResponseEntity<AccountAccessConsent> createConsent(@ApiParam(value = "Create Consent Request", required = true)
-                                                       @Valid
-                                                       @RequestBody CreateAccountAccessConsentRequest request);
+    ResponseEntity<InternationalStandingOrderConsent> createConsent(@ApiParam(value = "Create Consent Request", required = true)
+                                                                    @Valid
+                                                                    @RequestBody CreateInternationalStandingOrderConsentRequest request);
 
 
-    @ApiOperation(value = "Get Account Access Consent")
+    @ApiOperation(value = "Get International Standing Order Consent")
     @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "AccountAccessConsent object representing the consent created",
-                    response = AccountAccessConsent.class),
+            @ApiResponse(code = 200, message = "InternationalStandingOrderConsent object representing the consent created",
+                         response = InternationalStandingOrderConsent.class),
             @ApiResponse(code = 400, message = "Bad request", response = OBErrorResponse1.class),
             @ApiResponse(code = 403, message = "Forbidden", response = OBErrorResponse1.class),
             @ApiResponse(code = 404, message = "Not found"),
@@ -74,17 +70,17 @@ public interface AccountAccessConsentApi {
             @ApiResponse(code = 406, message = "Not Acceptable"),
             @ApiResponse(code = 500, message = "Internal Server Error", response = OBErrorResponse1.class)
     })
-    @RequestMapping(value = "/account-access-consents/{consentId}",
+    @RequestMapping(value = "/international-standing-order-consents/{consentId}",
             produces = {"application/json; charset=utf-8"},
             method = RequestMethod.GET)
-    ResponseEntity<AccountAccessConsent> getConsent(@PathVariable(value = "consentId") String consentId,
-                                                    @RequestHeader(value = "x-api-client-id") String apiClientId);
+    ResponseEntity<InternationalStandingOrderConsent> getConsent(@PathVariable(value = "consentId") String consentId,
+                                                                 @RequestHeader(value = "x-api-client-id") String apiClientId);
 
 
-    @ApiOperation(value = "Authorise Account Access Consent")
+    @ApiOperation(value = "Authorise International Standing Order Consent")
     @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "AccountAccessConsent object representing the consent created",
-                    response = AccountAccessConsent.class),
+            @ApiResponse(code = 200, message = "InternationalStandingOrderConsent object representing the consent created",
+                         response = InternationalStandingOrderConsent.class),
             @ApiResponse(code = 400, message = "Bad request", response = OBErrorResponse1.class),
             @ApiResponse(code = 403, message = "Forbidden", response = OBErrorResponse1.class),
             @ApiResponse(code = 404, message = "Not found"),
@@ -92,20 +88,20 @@ public interface AccountAccessConsentApi {
             @ApiResponse(code = 406, message = "Not Acceptable"),
             @ApiResponse(code = 500, message = "Internal Server Error", response = OBErrorResponse1.class)
     })
-    @RequestMapping(value = "/account-access-consents/{consentId}/authorise",
+    @RequestMapping(value = "/international-standing-order-consents/{consentId}/authorise",
             consumes = {"application/json; charset=utf-8"},
             produces = {"application/json; charset=utf-8"},
             method = RequestMethod.POST)
-    ResponseEntity<AccountAccessConsent> authoriseConsent(@PathVariable(value = "consentId") String consentId,
-                                                          @ApiParam(value = "Authorise Consent Request", required = true)
-                                                          @Valid
-                                                          @RequestBody AuthoriseAccountAccessConsentRequest request);
+    ResponseEntity<InternationalStandingOrderConsent> authoriseConsent(@PathVariable(value = "consentId") String consentId,
+                                                                       @ApiParam(value = "Authorise Consent Request", required = true)
+                                                                       @Valid
+                                                                       @RequestBody AuthorisePaymentConsentRequest request);
 
 
-    @ApiOperation(value = "Reject Account Access Consent")
+    @ApiOperation(value = "Reject International Standing Order Consent")
     @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "AccountAccessConsent object representing the consent created",
-                    response = AccountAccessConsent.class),
+            @ApiResponse(code = 200, message = "InternationalStandingOrderConsent object representing the consent created",
+                         response = InternationalStandingOrderConsent.class),
             @ApiResponse(code = 400, message = "Bad request", response = OBErrorResponse1.class),
             @ApiResponse(code = 403, message = "Forbidden", response = OBErrorResponse1.class),
             @ApiResponse(code = 404, message = "Not found"),
@@ -113,19 +109,21 @@ public interface AccountAccessConsentApi {
             @ApiResponse(code = 406, message = "Not Acceptable"),
             @ApiResponse(code = 500, message = "Internal Server Error", response = OBErrorResponse1.class)
     })
-    @RequestMapping(value = "/account-access-consents/{consentId}/reject",
+    @RequestMapping(value = "/international-standing-order-consents/{consentId}/reject",
             consumes = {"application/json; charset=utf-8"},
             produces = {"application/json; charset=utf-8"},
             method = RequestMethod.POST)
-    ResponseEntity<AccountAccessConsent> rejectConsent(@PathVariable(value = "consentId") String consentId,
-                                                       @ApiParam(value = "Reject Consent Request", required = true)
-                                                       @Valid
-                                                       @RequestBody RejectConsentRequest request);
+    ResponseEntity<InternationalStandingOrderConsent> rejectConsent(@PathVariable(value = "consentId") String consentId,
+                                                                    @ApiParam(value = "Reject Consent Request", required = true)
+                                                                    @Valid
+                                                                    @RequestBody RejectConsentRequest request);
 
 
-    @ApiOperation(value = "Delete Account Access Consent")
+
+    @ApiOperation(value = "Consume International Standing Order Consent")
     @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "Delete successful"),
+            @ApiResponse(code = 200, message = "InternationalStandingOrderConsent object representing the consent created",
+                         response = InternationalStandingOrderConsent.class),
             @ApiResponse(code = 400, message = "Bad request", response = OBErrorResponse1.class),
             @ApiResponse(code = 403, message = "Forbidden", response = OBErrorResponse1.class),
             @ApiResponse(code = 404, message = "Not found"),
@@ -133,10 +131,13 @@ public interface AccountAccessConsentApi {
             @ApiResponse(code = 406, message = "Not Acceptable"),
             @ApiResponse(code = 500, message = "Internal Server Error", response = OBErrorResponse1.class)
     })
-    @RequestMapping(value = "/account-access-consents/{consentId}",
-            method = RequestMethod.DELETE)
-    ResponseEntity<Void> deleteConsent(@PathVariable(value = "consentId") String consentId,
-                                       @RequestHeader(value = "x-api-client-id") String apiClientId);
-
+    @RequestMapping(value = "/international-standing-order-consents/{consentId}/consume",
+            consumes = {"application/json; charset=utf-8"},
+            produces = {"application/json; charset=utf-8"},
+            method = RequestMethod.POST)
+    ResponseEntity<InternationalStandingOrderConsent> consumeConsent(@PathVariable(value = "consentId") String consentId,
+                                                                     @ApiParam(value = "Consume Consent Request", required = true)
+                                                                     @Valid
+                                                                     @RequestBody ConsumePaymentConsentRequest request);
 
 }
