+++ secure-api-gateway-ob-uk-rcs-consent-store/secure-api-gateway-ob-uk-rcs-consent-store-client/src/main/java/com/forgerock/sapi/gateway/rcs/consent/store/client/customerinfo/v1_0/CustomerInfoConsentStoreClient.java
@@ -22,7 +22,7 @@ import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.customerinfo.v1_0.
 import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.customerinfo.v1_0.CustomerInfoConsent;
 
 /**
- * Client for interacting with com.forgerock.sapi.gateway.rcs.consent.store.api.customerinfo.v1_0.CustomerInfoConsentApi
+ * Client for interacting with com.forgerock.sapi.gateway.rcs.consent.store.api.customerinfo.CustomerInfoConsentApi
  */
 public interface CustomerInfoConsentStoreClient {
     CustomerInfoConsent createConsent(CreateCustomerInfoConsentRequest createConsentRequest) throws ConsentStoreClientException;
