+++ secure-api-gateway-ob-uk-rcs-consent-store/secure-api-gateway-ob-uk-rcs-consent-store-client/src/main/java/com/forgerock/sapi/gateway/rcs/consent/store/client/funds/v3_1_10/FundsConfirmationConsentStoreClient.java
@@ -22,7 +22,7 @@ import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.funds.v3_1_10.Crea
 import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.funds.v3_1_10.FundsConfirmationConsent;
 
 /**
- * Client for interacting with com.forgerock.sapi.gateway.rcs.consent.store.api.funds.v3_1_10.FundsConfirmationConsentApi
+ * Client for interacting with com.forgerock.sapi.gateway.rcs.consent.store.api.funds.FundsConfirmationConsentApi
  */
 public interface FundsConfirmationConsentStoreClient {
     FundsConfirmationConsent createConsent(CreateFundsConfirmationConsentRequest createConsentRequest) throws ConsentStoreClientException;
